# internal/handler/report.go
package handler

import (
	"net/http"
	"strconv"
	"time"

	"github.com/heartgryphon/dsp/internal/auth"
	"github.com/heartgryphon/dsp/internal/reporting"
)

// HandleCampaignStats godoc
// @Summary Get campaign stats
// @Tags reports
// @Security ApiKeyAuth
// @Produce json
// @Param id path int true "Campaign ID"
// @Param from query string false "Start date (YYYY-MM-DD)"
// @Param to query string false "End date (YYYY-MM-DD)"
// @Success 200 {object} reporting.CampaignStats
// @Router /reports/campaign/{id}/stats [get]
func (d *Deps) HandleCampaignStats(w http.ResponseWriter, r *http.Request) {
	if d.ReportStore == nil {
		WriteError(w, http.StatusServiceUnavailable, "ClickHouse not connected")
		return
	}
	id, err := strconv.ParseInt(r.PathValue("id"), 10, 64)
	if err != nil {
		WriteError(w, http.StatusBadRequest, "invalid id")
		return
	}
	from, to := ParseDateRange(r)
	stats, err := d.ReportStore.GetCampaignStats(r.Context(), uint64(id), from, to)
	if err != nil {
		WriteError(w, http.StatusInternalServerError, err.Error())
		return
	}
	WriteJSON(w, http.StatusOK, stats)
}

// HandleHourlyStats godoc
// @Summary Get hourly stats
// @Tags reports
// @Security ApiKeyAuth
// @Produce json
// @Param id path int true "Campaign ID"
// @Param date query string false "Date (YYYY-MM-DD)"
// @Success 200 {array} reporting.HourlyStats
// @Router /reports/campaign/{id}/hourly [get]
func (d *Deps) HandleHourlyStats(w http.ResponseWriter, r *http.Request) {
	if d.ReportStore == nil {
		WriteError(w, http.StatusServiceUnavailable, "ClickHouse not connected")
		return
	}
	id, err := strconv.ParseInt(r.PathValue("id"), 10, 64)
	if err != nil {
		WriteError(w, http.StatusBadRequest, "invalid id")
		return
	}
	dateStr := r.URL.Query().Get("date")
	date := time.Now().UTC()
	if dateStr != "" {
		if d, err := time.Parse("2006-01-02", dateStr); err == nil {
			date = d
		}
	}
	stats, err := d.ReportStore.GetHourlyStats(r.Context(), uint64(id), date)
	if err != nil {
		WriteError(w, http.StatusInternalServerError, err.Error())
		return
	}
	if stats == nil {
		stats = []reporting.HourlyStats{}
	}
	WriteJSON(w, http.StatusOK, stats)
}

// HandleGeoBreakdown godoc
// @Summary Get geo breakdown
// @Tags reports
// @Security ApiKeyAuth
// @Produce json
// @Param id path int true "Campaign ID"
// @Success 200 {array} reporting.GeoStats
// @Router /reports/campaign/{id}/geo [get]
func (d *Deps) HandleGeoBreakdown(w http.ResponseWriter, r *http.Request) {
	if d.ReportStore == nil {
		WriteError(w, http.StatusServiceUnavailable, "ClickHouse not connected")
		return
	}
	id, err := strconv.ParseInt(r.PathValue("id"), 10, 64)
	if err != nil {
		WriteError(w, http.StatusBadRequest, "invalid id")
		return
	}
	from, to := ParseDateRange(r)
	stats, err := d.ReportStore.GetGeoBreakdown(r.Context(), uint64(id), from, to)
	if err != nil {
		WriteError(w, http.StatusInternalServerError, err.Error())
		return
	}
	if stats == nil {
		stats = []reporting.GeoStats{}
	}
	WriteJSON(w, http.StatusOK, stats)
}

// HandleBidTransparency godoc
// @Summary Get bid-level details
// @Tags reports
// @Security ApiKeyAuth
// @Produce json
// @Param id path int true "Campaign ID"
// @Param limit query int false "Limit" default(100)
// @Param offset query int false "Offset" default(0)
// @Success 200 {array} reporting.BidDetail
// @Router /reports/campaign/{id}/bids [get]
func (d *Deps) HandleBidTransparency(w http.ResponseWriter, r *http.Request) {
	if d.ReportStore == nil {
		WriteError(w, http.StatusServiceUnavailable, "ClickHouse not connected")
		return
	}
	id, err := strconv.ParseInt(r.PathValue("id"), 10, 64)
	if err != nil {
		WriteError(w, http.StatusBadRequest, "invalid id")
		return
	}
	from, to := ParseDateRange(r)
	limit, _ := strconv.Atoi(r.URL.Query().Get("limit"))
	offset, _ := strconv.Atoi(r.URL.Query().Get("offset"))
	if limit <= 0 {
		limit = 50
	}
	bids, err := d.ReportStore.GetBidTransparency(r.Context(), uint64(id), from, to, limit, offset)
	if err != nil {
		WriteError(w, http.StatusInternalServerError, err.Error())
		return
	}
	if bids == nil {
		bids = []reporting.BidDetail{}
	}
	WriteJSON(w, http.StatusOK, bids)
}

// HandleAttribution godoc
// @Summary Get attribution report
// @Tags reports
// @Security ApiKeyAuth
// @Produce json
// @Param id path int true "Campaign ID"
// @Success 200 {object} reporting.AttributionReport
// @Router /reports/campaign/{id}/attribution [get]
func (d *Deps) HandleAttribution(w http.ResponseWriter, r *http.Request) {
	if d.ReportStore == nil {
		WriteError(w, http.StatusServiceUnavailable, "ClickHouse not connected")
		return
	}
	id, err := strconv.ParseInt(r.PathValue("id"), 10, 64)
	if err != nil {
		WriteError(w, http.StatusBadRequest, "invalid id")
		return
	}
	from, to := ParseDateRange(r)
	model := r.URL.Query().Get("model")
	if model == "" {
		model = "last_click"
	}
	limit, _ := strconv.Atoi(r.URL.Query().Get("limit"))

	report, err := d.ReportStore.GetAttributionReport(r.Context(), uint64(id), from, to, model, limit)
	if err != nil {
		WriteError(w, http.StatusInternalServerError, err.Error())
		return
	}
	WriteJSON(w, http.StatusOK, report)
}

// HandleOverviewStats godoc
// @Summary Get today overview stats
// @Tags reports
// @Security ApiKeyAuth
// @Produce json
// @Success 200 {object} object{today_spend_cents=integer,today_impressions=integer,today_clicks=integer,ctr=number,balance_cents=integer}
// @Router /reports/overview [get]
func (d *Deps) HandleOverviewStats(w http.ResponseWriter, r *http.Request) {
	if d.ReportStore == nil {
		WriteJSON(w, http.StatusOK, map[string]any{"today_spend_cents": 0})
		return
	}
	advID := auth.AdvertiserIDFromContext(r.Context())

	campaigns, err := d.Store.ListCampaigns(r.Context(), advID)
	if err != nil {
		WriteJSON(w, http.StatusOK, map[string]any{"today_spend_cents": 0})
		return
	}

	today := time.Now().UTC().Truncate(24 * time.Hour)
	tomorrow := today.Add(24 * time.Hour)
	var totalSpend uint64
	var totalImpressions, totalClicks uint64
	for _, c := range campaigns {
		stats, err := d.ReportStore.GetCampaignStats(r.Context(), uint64(c.ID), today, tomorrow)
		if err != nil || stats == nil {
			continue
		}
		totalSpend += stats.SpendCents
		totalImpressions += stats.Impressions
		totalClicks += stats.Clicks
	}

	var ctr float64
	if totalImpressions > 0 {
		ctr = float64(totalClicks) / float64(totalImpressions) * 100
	}

	// Get advertiser balance
	var balanceCents int64
	if d.BillingSvc != nil {
		bal, _, err := d.BillingSvc.GetBalance(r.Context(), advID)
		if err == nil {
			balanceCents = bal
		}
	}

	WriteJSON(w, http.StatusOK, map[string]any{
		"today_spend_cents": totalSpend,
		"today_impressions": totalImpressions,
		"today_clicks":      totalClicks,
		"ctr":               ctr,
		"balance_cents":     balanceCents,
	})
}